test(#64): 로비 목록 필터링 및 정렬 테스트 추가
